+++ package.json
@@ -43,7 +43,7 @@
     "react-native-screens": "~2.15.0",
     "react-native-unimodules": "~0.12.0",
     "react-native-web": "~0.13.12",
-    "react-navigation-drawer": "^2.6.0",
+    "react-navigation-drawer": "^2.7.0",
     "react-navigation-stack": "^2.10.2",
     "yup": "^0.32.9"
   },
