+++ routes/TeamDrawer.js
@@ -1,0 +1,9 @@
+import { createDrawerNavigator } from 'react-navigation-drawer';
+import { createAppContainer } from 'react-navigation';
+
+
+const TeamDrawerNavigation = createDrawerNavigator({
+
+
+
+})
