+++ src/screens/LoginRoutes/ProfileScreen/styles.js
@@ -1,5 +1,5 @@
 import {StyleSheet} from 'react-native';
-import Colors from '../styles';
+import Colors from '../../styles';
 
 export default StyleSheet.create({
