+++ src/screens/TeamRoutes/Members/index.js
@@ -1,0 +1,16 @@
+import React, {Component} from "react";
+import {View, Text} from 'react-native';
+
+
+export default class MemberPage extends Component{
+
+    render(){
+        return(
+            <View>
+                <Text>Student page</Text>
+            </View>
+        )
+
+    }
+
+}
