+++ src/screens/index.js
@@ -1,6 +1,6 @@
-export {default as LoginScreen} from './LoginScreen/LoginScreen';
-export {default as HomeScreen} from './HomeScreen/HomeScreen';
-export {default as CreateAccountScreen} from './CreateAccountScreen/CreateAccountScreen'
-export {default as ProfileScreen} from './ProfileScreen/ProfileScreen'
-export {default as CreateProfileScreen} from './CreateProfileScreen/CreateProfileScreen'
-export {default as Dashboard} from './DashboardScreen/DashboardScreen'
+export {default as LoginScreen} from './LoginRoutes/LoginScreen/LoginScreen';
+export {default as HomeScreen} from './LoginRoutes/HomeScreen/HomeScreen';
+export {default as CreateAccountScreen} from './LoginRoutes/CreateAccountScreen/CreateAccountScreen'
+export {default as ProfileScreen} from './LoginRoutes/ProfileScreen/ProfileScreen'
+export {default as CreateProfileScreen} from './LoginRoutes/CreateProfileScreen/CreateProfileScreen'
+export {default as Dashboard} from './TeamRoutes/DashboardScreen/DashboardScreen'
